Several minor changes to make the code x86_64 ready.

+++ src/d4/app/Application.d
@@ -360,7 +360,7 @@ protected:
       Option[] options = helpOptions();
       options.sort();
 
-      uint maxLength = 0;
+      size_t maxLength = 0;
       foreach ( option; options ) {
          maxLength = max( maxLength, option.name.length );
       }
